// src/shared/api/models/getApiProductsListing200ProductsItemProductsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiProductsListing200ProductsItemProductsItemDescription } from './getApiProductsListing200ProductsItemProductsItemDescription';
import type { GetApiProductsListing200ProductsItemProductsItemImage } from './getApiProductsListing200ProductsItemProductsItemImage';
import type { GetApiProductsListing200ProductsItemProductsItemSku } from './getApiProductsListing200ProductsItemProductsItemSku';
import type { GetApiProductsListing200ProductsItemProductsItemCategoryId } from './getApiProductsListing200ProductsItemProductsItemCategoryId';
import type { GetApiProductsListing200ProductsItemProductsItemSupplierId } from './getApiProductsListing200ProductsItemProductsItemSupplierId';
import type { GetApiProductsListing200ProductsItemProductsItemBrand } from './getApiProductsListing200ProductsItemProductsItemBrand';
import type { GetApiProductsListing200ProductsItemProductsItemOutletId } from './getApiProductsListing200ProductsItemProductsItemOutletId';

export type GetApiProductsListing200ProductsItemProductsItem = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    description: GetApiProductsListing200ProductsItemProductsItemDescription;
    image: GetApiProductsListing200ProductsItemProductsItemImage;
    sku: GetApiProductsListing200ProductsItemProductsItemSku;
    categoryId: GetApiProductsListing200ProductsItemProductsItemCategoryId;
    supplierId: GetApiProductsListing200ProductsItemProductsItemSupplierId;
    brand: GetApiProductsListing200ProductsItemProductsItemBrand;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    price: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    costPrice: number;
    isActive: boolean;
    availableOnline: boolean;
    outletId: GetApiProductsListing200ProductsItemProductsItemOutletId;
    createdAt: string;
    updatedAt: string;
    stock?: number;
    tax?: string[];
};