// src/shared/api/models/putApiUpdateProductsBodyDataItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiUpdateProductsBodyDataItemSku } from './putApiUpdateProductsBodyDataItemSku';
import type { PutApiUpdateProductsBodyDataItemCategoryId } from './putApiUpdateProductsBodyDataItemCategoryId';
import type { PutApiUpdateProductsBodyDataItemPrice } from './putApiUpdateProductsBodyDataItemPrice';
import type { PutApiUpdateProductsBodyDataItemCostPrice } from './putApiUpdateProductsBodyDataItemCostPrice';

export type PutApiUpdateProductsBodyDataItem = {
    id: string;
    name?: string;
    sku?: PutApiUpdateProductsBodyDataItemSku;
    categoryId?: PutApiUpdateProductsBodyDataItemCategoryId;
    price?: PutApiUpdateProductsBodyDataItemPrice;
    costPrice?: PutApiUpdateProductsBodyDataItemCostPrice;
};